Fix build errors from code review changes
- Fix narrowing warnings in alp_wrapper.cc (uint64_t -> int64_t casts
  for CompressionProgress/DecompressionProgress fields)
- Fix narrowing in alp.cc (best_compressed_size_bytes uint64_t -> int64_t)
- Fix remaining Decode() call sites in alp_test.cc to use new parameter
  order (num_elements, comp, comp_size, output)

+++ cpp/src/arrow/util/alp/alp.cc
@@ -664,8 +664,8 @@ AlpEncodingParameters AlpCompression<T>::CreateEncodingParameters(
     combinations.push_back(best_k_combinations[i].exponent_and_factor);
   }
 
-  const uint64_t best_compressed_size_bytes =
-      static_cast<uint64_t>(bit_util::BytesForBits(best_compressed_size_bits));
+  const int64_t best_compressed_size_bytes =
+      bit_util::BytesForBits(best_compressed_size_bits);
   return {combinations, best_compressed_size_bytes};
 }
